classificator_back/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.2.1
capstone==5.0.3
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.8
colored-traceback==0.4.2
cryptography==44.0.0
exceptiongroup==1.2.2
fastapi==0.115.7
greenlet==3.1.1
h11==0.14.0
idna==3.10
intervaltree==3.1.0
Mako==1.3.8
MarkupSafe==3.0.2
mem_edit==0.7
numpy==2.2.0
packaging==24.2
paramiko==3.5.0
passlib==1.7.4
plumbum==1.9.0
psutil==6.1.0
psycopg2-binary==2.9.10
pwntools==4.13.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyelftools==0.31
Pygments==2.18.0
PyJWT==2.10.1
PyNaCl==1.5.0
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
r2pipe==1.9.4
requests==2.32.3
ROPGadget==7.5
rpyc==6.0.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.37
SQLAlchemy-Utils==0.41.2
starlette==0.45.2
typing_extensions==4.12.2
unicorn==2.1.1
unix-ar==0.2.1
urllib3==2.2.3
uvicorn==0.34.0
zstandard==0.23.0
